< fix > fix bugs in distributer.cpp by adding the initialization of startz. < range > module_pw
